requirements.txt
astroid==3.0.2
attrs==23.1.0
aws-cdk-lib==2.114.0
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
cattrs==23.2.3
cdk-nag==2.27.222
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
constructs==10.3.0
dill==0.3.7
dparse==0.6.3
idna==3.6
importlib-resources==6.1.1
isort==5.13.2
jsii==1.93.0
mccabe==0.7.0
packaging==21.3
platformdirs==4.1.0
publication==0.0.3
pylint==3.0.3
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safety==2.3.5
six==1.16.0
tomlkit==0.12.3
typeguard==2.13.3
typing_extensions==4.9.0
urllib3==2.1.0
